指摘事項の修正
