+++ src/components/page/Top/Top.page.tsx
@@ -3,7 +3,7 @@ import type { Category } from '@/api/category/types';
 import type { Recipe } from '@/api/recipe/types';
 import style from '@/components/page/Top/Top.module.scss';
 import { Top } from '@/components/page/Top/Top';
-import { Footer } from '@/components/common/Footer';
+import { Footer } from '@/components/common/footer/Footer';
 
 interface Props {
   category: MicroCMSListResponse<Category>;
@@ -13,9 +13,7 @@ interface Props {
 export const TopPage = (props: Props): JSX.Element => {
   return (
     <>
-      <main>
-        <Top {...props} />
-      </main>
+      <Top {...props} />
       <Footer />
     </>
   );
